パッシブスキルと機能とアイテムに管理クラスを実装。管理クラスの持ち主からアクセスできる。新たに麻痺を実装。確率でアクティブスキル（魔法スキル含め）が失敗する。エイリアス宣言と前方宣言の書きミスを修正。

+++ Game/Game/Item/ItemManager.h
@@ -2,7 +2,7 @@
 
 namespace nsAWA {
 
-	//エイリアス宣言
+	//前方宣言
 	namespace nsItem {
 
 		class CItem;
@@ -58,7 +58,7 @@ namespace nsAWA {
 				if (m_selectItemNum < 0) {
 
 					//最後のアイテムを見る。
-					m_selectItemNum = m_itemList.size() - 1;
+					m_selectItemNum = static_cast<int>(m_itemList.size() - 1);
 				}
 			}
